Split up modal in different components
